refactor(atoms): inline TimezoneChip into EventTime
Move the single-use TimezoneChip composition into its only parent (EventTime)
as a private component, leaving Chip a pure presentational atom with no
date/tooltip logic. Chip's props type is now `ChipProps` (NextUI's import is
aliased to `NextUIChipProps`).

Refs #14
